refactor: rename data folder
* Rename into cache, as it will
contain implementations of the cache structures
